custom_components/visonic/examples/requirements.txt
aconsole==0.0.8
pyserial_asyncio==0.4
